refactor(config): extract config update service
