+++ src/agent/cmd/daemon/config_update_compat_test.go
@@ -1,0 +1,15 @@
+package main
+
+import "aiden-agent/internal/configupdate"
+
+type configUpdateResult = configupdate.Result
+
+const hasAPIKeyPlaceholder = "***"
+
+func updateConfigFile(path string, patch []byte) (configUpdateResult, error) {
+	return configupdate.NewService().Update(path, patch)
+}
+
+func boolPtr(value bool) *bool {
+	return &value
+}
